backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
web3==6.0.0
pymongo==4.3.3
motor==3.1.2
loguru==0.7.0
pytest==7.3.1
httpx==0.24.0
python-multipart==0.0.6
requests==2.28.2
tenacity==8.2.2
openai>=1.3.0
aiohttp==3.8.4